+++ src/main/java/blackjack/dto/DealerMatchDto.java
@@ -1,4 +1,4 @@
-package blackjack.domain.game.dto;
+package blackjack.dto;
 
 import blackjack.domain.game.Referee;
 import blackjack.domain.game.ResultCount;
