+++ src/main/java/blackjack/view/OutputView.java
@@ -4,11 +4,11 @@ import blackjack.domain.card.Card;
 import blackjack.domain.game.BlackjackGame;
 import blackjack.domain.game.ResultCount;
 import blackjack.domain.game.ResultType;
-import blackjack.domain.game.dto.DealerMatchDto;
-import blackjack.domain.game.dto.PlayerMatchDto;
 import blackjack.domain.participant.Dealer;
 import blackjack.domain.participant.Participant;
 import blackjack.domain.participant.Player;
+import blackjack.dto.DealerMatchDto;
+import blackjack.dto.PlayerMatchDto;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
